CORE: fixed missing cast for FourCC

+++ src/modules/core/Common.h
@@ -172,7 +172,7 @@ auto operator()(T t) const {
 };
 
 inline constexpr uint32_t FourCC(uint8_t a, uint8_t b, uint8_t c, uint8_t d) {
-	return ((uint32_t) (((d) << 24) | (uint32_t(c) << 16) | (uint32_t(b) << 8) | uint32_t(a)));
+	return ((uint32_t) ((uint32_t(d) << 24) | (uint32_t(c) << 16) | (uint32_t(b) << 8) | uint32_t(a)));
 }
 
 #define DIAG_STR(s) #s
